# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /denied-merges:
    get:
      operationId: listDeniedMerges
      tags: [deniedMerges]
      summary: List denied (ignored) ingredient-merge pairs
      description: >-
        Returns every denied merge pair — an unordered pair of ingredient names
        the user explicitly told the app to never propose merging together. The
        merge suggester filters these out of both AI and remembered suggestions.
        Available to any signed-in user.
      responses:
        "200":
          description: All denied merge pairs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeniedMergeList"
    post:
      operationId: saveDeniedMerges
      tags: [deniedMerges]
      summary: Save denied (ignored) ingredient-merge pairs
      description: >-
        Persists a batch of unordered name pairs the user denied. Pairs are
        normalized (trimmed, lowercased, sorted) and deduped; a pair that already
        exists is left as-is. Available to any signed-in user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveDeniedMergesInput"
      responses:
        "200":
          description: The full updated denied-pair list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeniedMergeList"
        "400":
          description: Invalid input
    delete:
      operationId: deleteDeniedMerges
      tags: [deniedMerges]
      summary: Un-deny (remove) ingredient-merge pairs
      description: >-
        Removes a batch of denied pairs so they may be suggested again. Pairs are
        matched case-insensitively regardless of order. Available to any
        signed-in user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveDeniedMergesInput"
      responses:
        "200":
          description: The full updated denied-pair list after removal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeniedMergeList"
        "400":
          description: Invalid input
  /production-rules:
    get:
      operationId: listProductionRules
      tags: [productionRules]
      summary: List manager-defined production rules
      description: >-
        Returns every factory-wide production rule. Rules are global policy (not
        part of the per-day sync payload). Reading is available to any signed-in
        user so both apps can evaluate them; editing is manager-only.
      responses:
        "200":
          description: All production rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionRuleList"
    post:
      operationId: saveProductionRules
      tags: [productionRules]
      summary: Create or update production rules (manager only)
      description: >-
        Upserts a batch of production rules by id. Each rule is normalized and
        validated server-side; malformed rules are dropped. Manager role required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveProductionRulesInput"
      responses:
        "200":
          description: The full updated rule list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionRuleList"
        "400":
          description: Invalid input
        "403":
          description: Manager role required
    delete:
      operationId: deleteProductionRules
      tags: [productionRules]
      summary: Delete production rules by id (manager only)
      description: >-
        Removes a batch of production rules by id. Manager role required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteProductionRulesInput"
      responses:
        "200":
          description: The full updated rule list after removal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionRuleList"
        "400":
          description: Invalid input
        "403":
          description: Manager role required
...
components:
  schemas:
...
    DeniedMerge:
      type: object
      description: >-
        An unordered pair of ingredient names the user told the app to never
        propose merging together (matched case-insensitively, either direction).
      required:
        - nameA
        - nameB
      properties:
        nameA:
          type: string
          description: One name of the denied pair
        nameB:
          type: string
          description: The other name of the denied pair
    DeniedMergeList:
      type: object
      required:
        - denied
      properties:
        denied:
          type: array
          items:
            $ref: "#/components/schemas/DeniedMerge"
    SaveDeniedMergesInput:
      type: object
      required:
        - pairs
      properties:
        pairs:
          type: array
          description: The batch of denied pairs to add or remove
          items:
            $ref: "#/components/schemas/DeniedMerge"
    ProductionRule:
      type: object
      description: >-
        A manager-defined factory-wide production rule. Flat shape: only the
        fields relevant to `type` are used. "flexible" rules warn; "strict" rules
        block starting a run.
      required:
        - id
        - name
        - type
        - enforcement
        - enabled
      properties:
        id:
          type: string
          description: Stable client-generated id
        name:
          type: string
          description: Human-readable rule name
        type:
          type: string
          enum: [required-field, numeric-range, sequence]
        enforcement:
          type: string
          enum: [flexible, strict]
        enabled:
          type: boolean
        field:
          type: string
          nullable: true
          description: Run field key (required-field, numeric-range)
        min:
          type: number
          nullable: true
          description: Inclusive lower bound (numeric-range)
        max:
          type: number
          nullable: true
          description: Inclusive upper bound (numeric-range)
        attribute:
          type: string
          nullable: true
          description: Run attribute key for the transition (sequence)
        before:
          type: string
          nullable: true
          description: Disallowed preceding attribute value (sequence)
        after:
          type: string
          nullable: true
          description: Disallowed following attribute value (sequence)
        bypass:
          type: array
          nullable: true
          description: >-
            Exceptions: when the current run matches any of these conditions the
            rule is waived entirely (no warning, no block). Applies to any type.
          items:
            $ref: "#/components/schemas/RuleBypassCondition"
        checklist:
          type: array
          nullable: true
          description: >-
            Exceptions: ordered step labels a manager attaches to a (strict)
            rule. When the rule is violated and not bypassed, the operator must
            acknowledge every step before Start unblocks for that run.
          items:
            type: string
    RuleBypassCondition:
      type: object
      description: >-
        A single bypass condition for a production rule. The rule is waived when
        the current run's `field` matches `value` (text fields case-insensitive,
        number fields by numeric equality).
      required:
        - field
        - value
      properties:
        field:
          type: string
          description: Run field key (from the shared RULE_FIELDS catalog)
        value:
          type: string
          description: Value under which the rule is waived
    ProductionRuleList:
      type: object
      required:
        - rules
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/ProductionRule"
    SaveProductionRulesInput:
      type: object
      required:
        - rules
      properties:
        rules:
          type: array
          description: The batch of rules to create or update (by id)
          items:
            $ref: "#/components/schemas/ProductionRule"
    DeleteProductionRulesInput:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          description: The ids of the rules to delete
          items:
            type: string
...